pages/__tests__/about.test.tsx
import { render } from '@testing-library/react';
import * as customHooks from '../../lib/custom-hooks';
import * as SeoTags from '../../components/seoTags';
import * as HeroSub from '../../components/heroSub';
import * as AboutSection from '../../components/aboutSection';
import About from '../about';

describe('<About />', () => {
  afterEach(() => {
    jest.restoreAllMocks();
  });

  it('should render expected components', async () => {
    // mock to prevent re-render of hero section
    jest.spyOn(customHooks, 'useMounted').mockReturnValue(true);

    const seoTagsSpy = jest.spyOn(SeoTags, 'default');
    const heroSubSpy = jest.spyOn(HeroSub, 'default');
    const aboutSectionSpy = jest.spyOn(AboutSection, 'default');

    render(<About />);

    expect(seoTagsSpy).toBeCalledTimes(1);

    // TODO: update SEO tags later
    expect(seoTagsSpy).toBeCalledWith(
      {
        path: '/about',
        title: 'About Me',
        description:
          'An introduction of myself - my passion, experiences and interests',
        imageUrl: '',
        imageWidth: 0,
        imageHeight: 0,
      },
      {}
    );

    expect(heroSubSpy).toBeCalledTimes(1);
    expect(aboutSectionSpy).toBeCalledTimes(1);
  });
});
